+++ tests/parser/functions/test_block_number.py
@@ -2,7 +2,7 @@
 def test_block_number(get_contract_with_gas_estimation, w3):
     block_number_code = """
 @public
-def block_number() -> int128:
+def block_number() -> uint256:
     return block.number
     """
     c = get_contract_with_gas_estimation(block_number_code)
